Move SpecUtils to be part of Metalware Gem's API
